+++ src/main/java/com/footstep/domain/users/dto/changeProfileInfo/ChangeNicknameInfo.java
@@ -1,0 +1,13 @@
+package com.footstep.domain.users.dto.changeProfileInfo;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ChangeNicknameInfo {
+
+    private String nickname;
+}
